+++ Runtime/Api/V1/Auth/ApiKeyAuthStrategy.cs
@@ -1,4 +1,5 @@
-﻿using System.Threading.Tasks;
+﻿using System.Threading;
+using System.Threading.Tasks;
 using ReadyPlayerMe.Data;
 using UnityEngine;
 
@@ -13,14 +14,14 @@ namespace ReadyPlayerMe.Api.V1
             _settings = Resources.Load<Settings>("ReadyPlayerMeSettings");
         }
         
-        public Task AddAuthToRequestAsync<T>(ApiRequest<T> request)
+        public Task AddAuthToRequestAsync<T>(ApiRequest<T> request, CancellationToken cancellationToken = default)
         {
             request.Headers["X-API-KEY"] = _settings.ApiKey;
 
             return Task.CompletedTask;
         }
 
-        public Task<bool> TryRefreshAsync<T>(ApiRequest<T> request)
+        public Task<bool> TryRefreshAsync<T>(ApiRequest<T> request, CancellationToken cancellationToken = default)
         {
             return Task.FromResult(false);
         }