+++ Samples~/AvatarCreator/Scripts/Constants.cs
@@ -1,0 +1,12 @@
+
+namespace ReadyPlayerMe.Samples.AvatarCreator
+{
+    public static class Constants 
+    {
+        public const string STYLE_ASSET_LABEL = "baseModel";
+        public const string SKELETON_DEFINITION_LABEL = "SkeletonDefinitionConfig";
+        public const string CHARACTER_STYLE_TEMPLATE_LABEL = "CharacterStyleTemplateConfig";
+        public const string STORED_CHARACTER_PREF = "RPM_Creator_Character";
+        public const string STORED_CHARACTER_STYLE_PREF = "RPM_Creator_Character_STYLE";
+    }
+}