+++ CHANGELOG.md
@@ -3,4 +3,22 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.1.0] - 2024.MM.DD
+## [0.2.0] - 2024.12.02
+
+## Updated
+- Cancellation tokens now added as optional parameter to web requests and other async methods
+
+## Added
+- New Avatar Creator sample scenes
+
+## [0.1.0] - 2024.27.09
+
+First release of the Ready Player Me Next Gen SDK
+
+## Added
+- Style Manager and Cache Creator editor windows
+- Web API wrapper classes
+- Avatar loading and creation functionality
+- Player Locomotion Sample
+- Basic UI Sample
+- Avatar Creator Sample